requirements.txt
fastapi
pydantic
langchain
langchain-community
openai
pinecone
httpx
langchain_openai
python-dotenv
pymongo
uvicorn[standard]
langsmith